Centralize link preview system as a core component
Move link preview fetching out of the Status plugin into a new
Idno\Core\LinkPreview component so any plugin can benefit from it:

- Create Idno/Core/LinkPreview.php extending Component, with:
  - extractFirstUrl() static helper for URL extraction
  - buildPreviewFromData() static helper for normalizing OG metadata
  - 'published' event listener that auto-enqueues preview fetching
    for any entity with a body containing a URL
  - 'linkpreview/fetch' async handler that fetches metadata via
    UnfurledUrl and caches it on the entity as link_preview
- Register LinkPreview in start.php alongside other core components
- Remove all link preview logic from Status plugin (Main.php event
  handler, Status.php queue call and extractFirstUrl method)

Any plugin can now use link previews by:
1. Having a text `body` property on its entity
2. Rendering entity/LinkPreviewCard in its view template when
   $entity->link_preview is populated
